+++ test/tests.js
@@ -1,3 +1,11 @@
+// Required for exposing test results to the Sauce Labs API.
+// Can be removed when the following issue is fixed:
+// https://github.com/axemclion/grunt-saucelabs/issues/84
+QUnit.done(function (details) {
+	window.global_test_results = details;
+});
+
+
 var lifecycle = {
 	teardown: function () {
 		$.cookie.defaults = {};
@@ -65,7 +73,7 @@ test('json = true', function () {
 	if ('JSON' in window) {
 		$.cookie.json = true;
 		$.cookie('c', { foo: 'bar' });
-		deepEqual($.cookie('c'), { foo: 'bar'}, 'should parse JSON');
+		deepEqual($.cookie('c'), { foo: 'bar' }, 'should parse JSON');
 	} else {
 		ok(true);
 	}
@@ -82,24 +90,44 @@ test('not existing with json = true', function () {
 	}
 });
 
-test('invalid JSON string with json = true', function () {
+test('string with json = true', function () {
 	expect(1);
-
 
 	if ('JSON' in window) {
 		$.cookie.json = true;
 		$.cookie('c', 'v');
+		strictEqual($.cookie('c'), 'v', 'should return value');
+	} else {
+		ok(true);
+	}
+});
+
+test('invalid JSON string with json = true', function () {
+	expect(1);
+
+	if ('JSON' in window) {
+		$.cookie('c', 'v');
+		$.cookie.json = true;
 		strictEqual($.cookie('c'), undefined, "won't throw exception, returns undefined");
 	} else {
 		ok(true);
 	}
 });
 
-asyncTest('malformed cookie value in IE (#88, #117)', function() {
+test('invalid URL encoding', function () {
+	expect(1);
+	document.cookie = 'bad=foo%';
+	strictEqual($.cookie('bad'), undefined, "won't throw exception, returns undefined");
+	// Delete manually here because it requires raw === true...
+	$.cookie.raw = true;
+	$.removeCookie('bad');
+});
+
+asyncTest('malformed cookie value in IE (#88, #117)', function () {
 	expect(1);
 	// Sandbox in an iframe so that we can poke around with document.cookie.
 	var iframe = $('<iframe src="malformed_cookie.html"></iframe>')[0];
-	$(iframe).on('load', function() {
+	$(iframe).on('load', function () {
 		start();
 		if (iframe.contentWindow.ok) {
 			strictEqual(iframe.contentWindow.testValue, 'two', 'reads all cookie values, skipping duplicate occurences of "; "');
@@ -113,20 +141,30 @@ asyncTest('malformed cookie value in IE (#88, #117)', function() {
 	document.body.appendChild(iframe);
 });
 
-test('call without arguments', function() {
+test('Call to read all when there are cookies', function () {
 	$.cookie('c', 'v');
 	$.cookie('foo', 'bar');
-	deepEqual($.cookie(), {
-		c: 'v',
-		foo: 'bar'
-	}, 'should return all cookies');
-	$.each($.cookie(), $.removeCookie);
+	deepEqual($.cookie(), { c: 'v', foo: 'bar' }, 'returns object containing all cookies');
+});
 
+test('Call to read all when there are no cookies at all', function () {
+	deepEqual($.cookie(), {}, 'returns empty object');
+});
+
+test('Call to read all with json: true', function () {
 	$.cookie.json = true;
 	$.cookie('c', { foo: 'bar' });
-	deepEqual($.cookie(), {
-		c: { foo: 'bar' }
-	}, 'returns all cookies with JSON parsed');
+	deepEqual($.cookie(), { c: { foo: 'bar' } }, 'returns JSON parsed cookies');
+});
+
+test('Call to read all with a badly encoded cookie', function () {
+	expect(1);
+	document.cookie = 'bad=foo%';
+	document.cookie = 'good=foo';
+	deepEqual($.cookie(), { good: 'foo' }, 'returns object containing all decodable cookies');
+	// Delete manually here because it requires raw === true...
+	$.cookie.raw = true;
+	$.removeCookie('bad');
 });
 
 
@@ -156,7 +194,19 @@ test('number', function () {
 	strictEqual($.cookie('c'), '1234', 'should write value');
 });
 
-test('expires option as days from now', function() {
+test('null', function () {
+	expect(1);
+	$.cookie('c', null);
+	strictEqual($.cookie('c'), 'null', 'should write value');
+});
+
+test('undefined', function () {
+	expect(1);
+	$.cookie('c', undefined);
+	strictEqual($.cookie('c'), 'undefined', 'should write value');
+});
+
+test('expires option as days from now', function () {
 	expect(1);
 	var sevenDaysFromNow = new Date();
 	sevenDaysFromNow.setDate(sevenDaysFromNow.getDate() + 7);
@@ -164,7 +214,19 @@ test('expires option as days from now', function() {
 		'should write the cookie string with expires');
 });
 
-test('expires option as Date instance', function() {
+test('expires option as fraction of a day', function () {
+	expect(1);
+
+	var now = new Date().getTime();
+	var expires = Date.parse($.cookie('c', 'v', { expires: 0.5 }).replace(/.+expires=/, ''));
+
+	// When we were using Date.setDate() fractions have been ignored
+	// and expires resulted in the current date. Allow 1000 milliseconds
+	// difference for execution time.
+	ok(expires > now + 1000, 'should write expires attribute with the correct date');
+});
+
+test('expires option as Date instance', function () {
 	expect(1);
 	var sevenDaysFromNow = new Date();
 	sevenDaysFromNow.setDate(sevenDaysFromNow.getDate() + 7);
@@ -188,7 +250,8 @@ test('raw = true', function () {
 	expect(1);
 	$.cookie.raw = true;
 	strictEqual($.cookie('c[1]', 'v[1]'), 'c[1]=v[1]', 'should not encode');
-	$.each($.cookie(), $.removeCookie);
+	// Delete manually here because it requires raw === true...
+	$.removeCookie('c[1]');
 });
 
 test('json = true', function () {
@@ -206,49 +269,53 @@ test('json = true', function () {
 
 module('removeCookie', lifecycle);
 
-test('deletion', function() {
+test('deletion', function () {
 	expect(1);
 	$.cookie('c', 'v');
 	$.removeCookie('c');
 	strictEqual(document.cookie, '', 'should delete the cookie');
 });
 
-test('return', function() {
-	expect(2);
-	strictEqual($.removeCookie('c'), false, "return false if the cookie wasn't found");
-
+test('when sucessfully deleted', function () {
+	expect(1);
 	$.cookie('c', 'v');
-	strictEqual($.removeCookie('c'), true, 'return true if the cookie was found');
+	strictEqual($.removeCookie('c'), true, 'returns true');
 });
 
-test('with options', function() {
-	expect(3);
-	var originalCookie = $.cookie;
-	var callCount = 0;
+test('when deletion failed', function () {
+	expect(1);
+	$.cookie('c', 'v');
 
-	$.cookie = function() {
-		callCount++;
-		if (callCount === 1)  {
-			// see https://github.com/carhartl/jquery-cookie/issues/99
-			strictEqual(arguments.length, 1, 'look up cookie instead of accidently writing a new');
-			return 'cookie'; // act as if a cookie was found...
-		}
-		if (callCount === 2) {
-			strictEqual(arguments[2].foo, 'bar', 'pass along options when deleting cookie');
+	var originalCookie = $.cookie;
+	$.cookie = function () {
+		// Stub deletion...
+		if (arguments.length === 1) {
+			return originalCookie.apply(null, arguments);
 		}
 	};
 
-	$.removeCookie('c', { foo: 'bar' });
-	strictEqual(callCount, 2);
+	strictEqual($.removeCookie('c'), false, 'returns false');
 
 	$.cookie = originalCookie;
 });
 
-test('passing options reference', function() {
+test('when cookie does not exist', function () {
+	expect(1);
+	strictEqual($.removeCookie('c'), false, 'returns false');
+});
+
+test('with options', function () {
 	expect(1);
 	var options = { path: '/' };
 	$.cookie('c', 'v', options);
+	$.removeCookie('c', options);
+	strictEqual(document.cookie, '', 'should delete the cookie');
+});
 
+test('passing options reference', function () {
+	expect(1);
+	var options = { path: '/' };
+	$.cookie('c', 'v', options);
 	$.removeCookie('c', options);
 	deepEqual(options, { path: '/' }, "won't alter options object");
 });
@@ -259,4 +326,20 @@ test('[] used in name', function () {
 	document.cookie = 'c[1]=foo';
 	$.removeCookie('c[1]');
 	strictEqual(document.cookie, '', 'delete the cookie');
+});
+
+
+module('conversion', lifecycle);
+
+test('read converter', function() {
+	expect(1);
+	$.cookie('c', '1');
+	strictEqual($.cookie('c', Number), 1, 'converts read value');
+});
+
+test('read converter with raw = true', function() {
+	expect(1);
+	$.cookie.raw = true;
+	$.cookie('c', '1');
+	strictEqual($.cookie('c', Number), 1, 'does not decode, but converts read value');
 });